refactor: move named to namedhist
