+++ src/hist/namedhist.py
